Additonal Tools for parsing SVG files
I included two other tools. The first tool includes a zip file of Python
scripts for parsing both the 2011 and 2012 SVG files. Since the Java XML
parser would crash when loading a SVG file, I resorted to Python for
parsing these files in order to generate the corresponding "_rooms.txt"
file. The other zip file contains a matlab script for visualizing the
"_rooms.txt" file to ensure the parsing of the SVG file was correct.